[Dsl] Where clause in Matmul.
Handle where clause for GEMM. This allows us to handle contraction operations
between a second-order tensor and a third-order tensor as matrix
multiplication, is some cases. The Tablegen entry generated has been extended
with the Dim field. The dim field tells how many dimensions are involved for
the computation of m, n, or k in the blas call.